+++ test/unit/Jamfile
@@ -44,6 +44,7 @@ local SOURCES =
     parse_path.cpp
     parse_query.cpp
     pct_string_view.cpp
+    router.cpp
     scheme.cpp
     segments_base.cpp
     segments_encoded_base.cpp
@@ -99,7 +100,6 @@ local SOURCES =
     rfc/unreserved_chars.cpp
     rfc/uri_rule.cpp
     rfc/uri_reference_rule.cpp
-    experimental/router.cpp
     ;
 for local f in $(SOURCES)
 {
